minor(history): split `HistoryListener`
