+++ src/EventListener/AddMonitorStampListener.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * This file is part of the zenstruck/messenger-monitor-bundle package.
+ *
+ * (c) Kevin Bond <kevinbond@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Zenstruck\Messenger\Monitor\EventListener;
+
+use Symfony\Component\Messenger\Event\SendMessageToTransportsEvent;
+use Zenstruck\Messenger\Monitor\Stamp\MonitorStamp;
+
+/**
+ * @author Kevin Bond <kevinbond@gmail.com>
+ *
+ * @internal
+ */
+final class AddMonitorStampListener
+{
+    public function __invoke(SendMessageToTransportsEvent $event): void
+    {
+        $event->setEnvelope($event->getEnvelope()->with(new MonitorStamp()));
+    }
+}
