+++ tests/Unit/EventListener/ReceiveMonitorStampListenerTest.php
@@ -9,22 +9,16 @@
  * file that was distributed with this source code.
  */
 
-namespace Zenstruck\Messenger\Monitor\Tests\Unit\History;
+namespace Zenstruck\Messenger\Monitor\Tests\Unit\EventListener;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Messenger\Envelope;
-use Symfony\Component\Messenger\Event\SendMessageToTransportsEvent;
-use Symfony\Component\Messenger\Event\WorkerMessageFailedEvent;
-use Symfony\Component\Messenger\Event\WorkerMessageHandledEvent;
 use Symfony\Component\Messenger\Event\WorkerMessageReceivedEvent;
 use Symfony\Component\Messenger\Stamp\HandledStamp;
 use Symfony\Component\Scheduler\Generator\MessageContext;
 use Symfony\Component\Scheduler\Messenger\ScheduledStamp;
 use Symfony\Component\Scheduler\Trigger\TriggerInterface;
-use Zenstruck\Messenger\Monitor\History\HistoryListener;
-use Zenstruck\Messenger\Monitor\History\Model\Results;
-use Zenstruck\Messenger\Monitor\History\ResultNormalizer;
-use Zenstruck\Messenger\Monitor\History\Storage;
+use Zenstruck\Messenger\Monitor\EventListener\ReceiveMonitorStampListener;
 use Zenstruck\Messenger\Monitor\Stamp\DisableMonitoringStamp;
 use Zenstruck\Messenger\Monitor\Stamp\MonitorStamp;
 use Zenstruck\Messenger\Monitor\Stamp\TagStamp;
@@ -32,36 +26,20 @@ use Zenstruck\Messenger\Monitor\Stamp\TagStamp;
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class HistoryListenerTest extends TestCase
+final class ReceiveMonitorStampListenerTest extends TestCase
 {
-    /**
-     * @test
-     */
-    public function adds_monitor_stamp(): void
-    {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
-        $envelope = new Envelope(new \stdClass());
-        $event = new SendMessageToTransportsEvent($envelope, []);
-
-        $this->assertNull($event->getEnvelope()->last(MonitorStamp::class));
-
-        $listener->addMonitorStamp($event);
-
-        $this->assertInstanceOf(MonitorStamp::class, $event->getEnvelope()->last(MonitorStamp::class));
-    }
-
     /**
      * @test
      */
     public function skips_standard_messages_without_monitor_stamp(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(new \stdClass());
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
         $this->assertEmpty($event->getEnvelope()->all(MonitorStamp::class));
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertEmpty($event->getEnvelope()->all(MonitorStamp::class));
     }
@@ -71,13 +49,13 @@ final class HistoryListenerTest extends TestCase
      */
     public function marks_standard_message_as_received(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(new \stdClass(), [new MonitorStamp()]);
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
         $this->assertFalse($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertTrue($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
         $this->assertSame('foo', $event->getEnvelope()->last(MonitorStamp::class)->transport());
@@ -89,7 +67,7 @@ final class HistoryListenerTest extends TestCase
      */
     public function marks_scheduled_message_as_received(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(new \stdClass(), [new ScheduledStamp(new MessageContext(
             'default',
             'id',
@@ -100,7 +78,7 @@ final class HistoryListenerTest extends TestCase
 
         $this->assertNull($event->getEnvelope()->last(MonitorStamp::class));
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertTrue($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
         $this->assertSame('foo', $event->getEnvelope()->last(MonitorStamp::class)->transport());
@@ -111,84 +89,13 @@ final class HistoryListenerTest extends TestCase
     /**
      * @test
      */
-    public function handles_success(): void
-    {
-        $envelope = new Envelope(new \stdClass(), [
-            (new MonitorStamp())->markReceived('foo'),
-            new HandledStamp('handler', 'return'),
-        ]);
-        $storage = $this->createMock(Storage::class);
-        $storage->expects($this->once())->method('save')->with(
-            $this->isInstanceOf(Envelope::class),
-            $this->isInstanceOf(Results::class),
-        );
-
-        $listener = new HistoryListener($storage, new ResultNormalizer(__DIR__), []);
-
-        $listener->handleSuccess($event = new WorkerMessageHandledEvent($envelope, 'foo'));
-
-        $this->assertTrue($event->getEnvelope()->last(MonitorStamp::class)->isFinished());
-    }
-
-    /**
-     * @test
-     */
-    public function handles_success_invalid(): void
-    {
-        $storage = $this->createMock(Storage::class);
-        $storage->expects($this->never())->method('save');
-
-        $listener = new HistoryListener($storage, new ResultNormalizer(__DIR__), []);
-
-        $listener->handleSuccess(new WorkerMessageHandledEvent(new Envelope(new \stdClass()), 'foo'));
-        $listener->handleSuccess(new WorkerMessageHandledEvent(new Envelope(new \stdClass(), [new MonitorStamp()]), 'foo'));
-    }
-
-    /**
-     * @test
-     */
-    public function handles_failure(): void
-    {
-        $envelope = new Envelope(new \stdClass(), [(new MonitorStamp())->markReceived('foo')]);
-        $exception = new \RuntimeException();
-        $storage = $this->createMock(Storage::class);
-        $storage->expects($this->once())->method('save')->with(
-            $this->isInstanceOf(Envelope::class),
-            $this->isInstanceOf(Results::class),
-            $exception,
-        );
-
-        $listener = new HistoryListener($storage, new ResultNormalizer(__DIR__), []);
-
-        $listener->handleFailure($event = new WorkerMessageFailedEvent($envelope, 'foo', $exception));
-
-        $this->assertTrue($event->getEnvelope()->last(MonitorStamp::class)->isFinished());
-    }
-
-    /**
-     * @test
-     */
-    public function handles_failure_invalid(): void
-    {
-        $storage = $this->createMock(Storage::class);
-        $storage->expects($this->never())->method('save');
-
-        $listener = new HistoryListener($storage, new ResultNormalizer(__DIR__), []);
-
-        $listener->handleFailure(new WorkerMessageFailedEvent(new Envelope(new \stdClass()), 'foo', new \RuntimeException()));
-        $listener->handleFailure(new WorkerMessageFailedEvent(new Envelope(new \stdClass(), [new MonitorStamp()]), 'foo', new \RuntimeException()));
-    }
-
-    /**
-     * @test
-     */
     public function can_disable_monitoring_with_envelope_stamp(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(new \stdClass(), [new MonitorStamp(), new DisableMonitoringStamp()]);
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertFalse($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
@@ -198,11 +105,11 @@ final class HistoryListenerTest extends TestCase
      */
     public function can_disable_monitoring_message_attribute(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(new DisabledMonitoringMessage(), [new MonitorStamp()]);
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertFalse($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
@@ -212,11 +119,11 @@ final class HistoryListenerTest extends TestCase
      */
     public function can_disable_monitoring_message_interface_attribute(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(new DisableMonitoringViaInterface(), [new MonitorStamp()]);
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertFalse($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
@@ -226,11 +133,11 @@ final class HistoryListenerTest extends TestCase
      */
     public function can_disable_monitoring_message_attribute_without_handler(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(new DisabledMonitoringWithoutHandlerMessage(), [new MonitorStamp()]);
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertFalse($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
@@ -240,14 +147,14 @@ final class HistoryListenerTest extends TestCase
      */
     public function can_disable_monitoring_message_without_handler(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(new \stdClass(), [
             new MonitorStamp(),
             new DisableMonitoringStamp(true),
         ]);
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertFalse($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
@@ -257,7 +164,7 @@ final class HistoryListenerTest extends TestCase
      */
     public function handle_disable_monitoring_message_attribute_with_handler(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(
             new EnabledMonitoringWithHandlerMessage(),
             [
@@ -267,7 +174,7 @@ final class HistoryListenerTest extends TestCase
         );
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertTrue($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
@@ -277,7 +184,7 @@ final class HistoryListenerTest extends TestCase
      */
     public function handle_disable_monitoring_message_with_handler(): void
     {
-        $listener = new HistoryListener($this->createMock(Storage::class), new ResultNormalizer(__DIR__), []);
+        $listener = new ReceiveMonitorStampListener([]);
         $envelope = new Envelope(
             new \stdClass(),
             [
@@ -288,7 +195,7 @@ final class HistoryListenerTest extends TestCase
         );
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertTrue($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
@@ -298,9 +205,7 @@ final class HistoryListenerTest extends TestCase
      */
     public function can_disable_monitoring_message_via_config(): void
     {
-        $listener = new HistoryListener(
-            $this->createMock(Storage::class),
-            new ResultNormalizer(__DIR__),
+        $listener = new ReceiveMonitorStampListener(
             [
                 MessageToDisableViaConfig::class,
             ]
@@ -308,7 +213,7 @@ final class HistoryListenerTest extends TestCase
         $envelope = new Envelope(new MessageToDisableViaConfig(), [new MonitorStamp()]);
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertFalse($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
@@ -318,9 +223,7 @@ final class HistoryListenerTest extends TestCase
      */
     public function can_disable_extended_monitoring_message_via_config(): void
     {
-        $listener = new HistoryListener(
-            $this->createMock(Storage::class),
-            new ResultNormalizer(__DIR__),
+        $listener = new ReceiveMonitorStampListener(
             [
                 MessageToDisableViaConfig::class,
             ]
@@ -328,7 +231,7 @@ final class HistoryListenerTest extends TestCase
         $envelope = new Envelope(new ExtendedMessageToDisableViaConfig(), [new MonitorStamp()]);
         $event = new WorkerMessageReceivedEvent($envelope, 'foo');
 
-        $listener->receiveMessage($event);
+        $listener->__invoke($event);
 
         $this->assertFalse($event->getEnvelope()->last(MonitorStamp::class)->isReceived());
     }
